Rename formatter to string_formatter
